+++ src/components/map/components/current-location/test/e2e/current-location.js
